# Make Wasm spec testsuite runnable via `miri`