--- a/crates/wasmi/tests/spec/run.rs
+++ b/crates/wasmi/tests/spec/run.rs
@@ -17,8 +17,8 @@ use wast::{
 };
 
 /// Runs the Wasm test spec identified by the given name.
-pub fn run_wasm_spec_test(name: &str, config: Config) {
-    let test = TestDescriptor::new(name);
+pub fn run_wasm_spec_test(name: &str, file: &str, config: Config) {
+    let test = TestDescriptor::new(name, file);
     let mut context = TestContext::new(&test, config);
 
     let mut lexer = Lexer::new(test.file());