The only thing preventing `miri` from working was usage of the filesystem which we were able to eliminate by usage of the `include_str` macro.

Local compile time benchmarks signaled no significant change:
```shell
cargo clean && time cargo build --test spec_shim
```

- **Before:** `50.11s user 3.71s system 651% cpu 8.259 total`
- **After:** `49.74s user 3.64s system 652% cpu 8.179 total`

Running tests time also has not changed significantly:
```shell
cargo clean && time cargo test --test spec_shim
```

- **Before:** `11.36s user 0.10s system 339% cpu 3.378 total`
- **After:** `11.26s user 0.09s system 338% cpu 3.357 total`